requirements.txt
mcp

google-genai 
google
python-dotenv
logging
dataclasses